projects/requirements.txt
alembic==1.5.7
alembic-utils==0.5.6
aniso8601==7.0.0
attrs==19.1.0
Click==7.0
Flask==1.0.3
Flask-Cors==3.0.8
Flask-Migrate==2.5.2
flask-restplus==0.12.1
Flask-SQLAlchemy==2.4.0
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
Mako==1.0.12
MarkupSafe==1.1.1
psycopg2==2.8.3
pyrsistent==0.15.2
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.1
six==1.12.0
SQLAlchemy==1.3.4
viztracer==0.12.3
Werkzeug==0.15.5
gunicorn==19.9.0
sqlalchemy_utils==0.37.4
